Changed week method behavior
Changed week.kid template to display a different layout which I will
eventually be a list of all weeks that have stats. I still need to modify
the Week model and change the template some more. But for now, if you try
to view the week page without passing a num parameter via get, it will show
"Select a Week"